front back align using ai

+++ backend/models/cart.go
@@ -15,7 +15,7 @@ type CartItem struct {
 	EntityName string    `json:"entityName,omitempty" bson:"entityName,omitempty"`
 	EntityType string    `json:"entityType,omitempty" bson:"entityType,omitempty"`
 	Quantity   int       `json:"quantity" bson:"quantity"`
-	Price      float64   `json:"price,omitempty" bson:"price,omitempty"`
+	Price      int64     `json:"price,omitempty" bson:"price,omitempty"` // CRITICAL FIX: Changed from float64 to int64 (stored in paise)
 	AddedAt    time.Time `json:"addedAt" bson:"addedAt"`
 }
 
@@ -24,11 +24,11 @@ type CheckoutSession struct {
 	UserID         string                `json:"userId" bson:"userId"`
 	Items          map[string][]CartItem `json:"items" bson:"items"`
 	Address        string                `json:"address" bson:"address"`
-	Total          float64               `json:"total" bson:"total"`
-	Subtotal       float64               `json:"subtotal" bson:"subtotal"`
-	Tax            float64               `json:"tax" bson:"tax"`
-	Delivery       float64               `json:"delivery" bson:"delivery"`
-	Discount       float64               `json:"discount" bson:"discount"`
+	Total          int64                 `json:"total" bson:"total"`       // CRITICAL FIX: Changed from float64 to int64 (stored in paise)
+	Subtotal       int64                 `json:"subtotal" bson:"subtotal"` // CRITICAL FIX: Changed from float64 to int64 (stored in paise)
+	Tax            int64                 `json:"tax" bson:"tax"`           // CRITICAL FIX: Changed from float64 to int64 (stored in paise)
+	Delivery       int64                 `json:"delivery" bson:"delivery"` // CRITICAL FIX: Changed from float64 to int64 (stored in paise)
+	Discount       int64                 `json:"discount" bson:"discount"` // CRITICAL FIX: Changed from float64 to int64 (stored in paise)
 	PaymentMethod  string                `json:"paymentMethod" bson:"paymentMethod"`
 	PaymentDetails interface{}           `json:"paymentDetails" bson:"paymentDetails"`
 	CreatedAt      time.Time             `json:"createdAt" bson:"createdAt"`
@@ -41,7 +41,7 @@ type Order struct {
 	Items         map[string][]CartItem `json:"items" bson:"items"` // grouped by category
 	Address       string                `json:"address" bson:"address"`
 	PaymentMethod string                `json:"paymentMethod" bson:"paymentMethod"`
-	Total         float64               `json:"total" bson:"total"`
+	Total         int64                 `json:"total" bson:"total"`   // CRITICAL FIX: Changed from float64 to int64 (stored in paise)
 	Status        string                `json:"status" bson:"status"` // e.g. "pending", "completed"
 	ApprovedBy    []string              `json:"approvedBy" bson:"approvedBy"`
 	CreatedAt     time.Time             `json:"createdAt" bson:"createdAt"`
